agents/fact_verifier_agent.py
from crewai import Agent, LLM
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

llm = LLM(
    model=os.getenv("OPENAI_MODEL_NAME"),
    temperature=0
)

verifier_agent = Agent(
    role='Fact Verification Expert',
    goal='Analyze search results and input analysis to determine if statements are true, false, or if there is not enough information. ',
    backstory="""You are an expert fact checker with a strong analytical mindset. 
    Your job is to carefully examine provided information and make accurate determinations 
    about the truthfulness of statements based solely on the evidence presented. You never make assumptions, only use the provided information and your answers are only based on the provided information. You cannot make searches with the tools, so you will answer based only on the provided information.""",
    verbose=True,
    allow_delegation=False,
    tools=[],  # This agent uses only the information provided by other agents,
    llm=llm
)